feat: Add iconsax-react and react-leaflet dependencies
This commit adds the iconsax-react and react-leaflet dependencies to the project. These libraries are needed for implementing new features and enhancing the user interface. The iconsax-react library provides a collection of customizable icons, while the react-leaflet library enables the integration of interactive maps into the application. These additions improve the functionality and visual appeal of the project.