feat: Add reverse geocoding functionality to MapModal component
This commit adds the ability to reverse geocode the selected location coordinates in the MapModal component. It integrates the axios library to make an HTTP request to the Nominatim API and retrieve the corresponding address for the coordinates. The retrieved address is then stored and displayed in the popup of the selected location marker on the map. This enhancement improves the user experience by providing a readable address for the selected location, enhancing the overall usability of the MapModal component.