feat: Add qrcode.react npm dependency
This commit adds the qrcode.react npm dependency to the project, with a version of 4.0.1. This package allows for the generation of QR codes in React applications. The addition of this dependency enhances the functionality of the application by enabling the generation and display of QR codes for various purposes.

Co-Authored-By: Victor olabisi <oguntayovictor65@gmail.com>