feat: Update package dependencies to include react-qr-code for generating QR codes
This commit updates the package dependencies to include the react-qr-code package with a version of 2.0.15. The react-qr-code package will be used for generating QR codes in the application, replacing the qrcode.react package. This change enhances the functionality of the application by improving the generation and display of QR codes.

Co-Authored-By: Victor olabisi <oguntayovictor65@gmail.com>