feat: Update MapModal and QRCodeModal components
This commit updates the MapModal component to include a new useEffect hook that centers the map on the selected location. It also updates the QRCodeModal component to use the QRCodeSVG component from the react-qr-code package instead of the QRCodeCanvas component. These changes enhance the functionality and improve the user experience of the application.

Co-Authored-By: Victor olabisi <oguntayovictor65@gmail.com>